+++ elixir/apps/domain/lib/domain/auth/identity/query.ex
@@ -87,7 +87,7 @@ defmodule Domain.Auth.Identity.Query do
   end
 
   def by_id_or_provider_identifier(queryable, id_or_provider_identifier) do
-    if Domain.Validator.valid_uuid?(id_or_provider_identifier) do
+    if Domain.Repo.valid_uuid?(id_or_provider_identifier) do
       where(
         queryable,
         [identities: identities],
