+++ elixir/apps/domain/lib/domain/config/definitions.ex
@@ -147,8 +147,8 @@ defmodule Domain.Config.Definitions do
     default: nil,
     changeset: fn changeset, key ->
       changeset
-      |> Domain.Validator.validate_uri(key, require_trailing_slash: true)
-      |> Domain.Validator.normalize_url(key)
+      |> Domain.Repo.Changeset.validate_uri(key, require_trailing_slash: true)
+      |> Domain.Repo.Changeset.normalize_url(key)
     end
   )
 
@@ -370,7 +370,7 @@ defmodule Domain.Config.Definitions do
   """
   defconfig(:tokens_key_base, :string,
     sensitive: true,
-    changeset: &Domain.Validator.validate_base64/2
+    changeset: &Domain.Repo.Changeset.validate_base64/2
   )
 
   @doc """
@@ -378,7 +378,7 @@ defmodule Domain.Config.Definitions do
   """
   defconfig(:tokens_salt, :string,
     sensitive: true,
-    changeset: &Domain.Validator.validate_base64/2
+    changeset: &Domain.Repo.Changeset.validate_base64/2
   )
 
   @doc """
@@ -386,7 +386,7 @@ defmodule Domain.Config.Definitions do
   """
   defconfig(:secret_key_base, :string,
     sensitive: true,
-    changeset: &Domain.Validator.validate_base64/2
+    changeset: &Domain.Repo.Changeset.validate_base64/2
   )
 
   @doc """
@@ -394,7 +394,7 @@ defmodule Domain.Config.Definitions do
   """
   defconfig(:live_view_signing_salt, :string,
     sensitive: true,
-    changeset: &Domain.Validator.validate_base64/2
+    changeset: &Domain.Repo.Changeset.validate_base64/2
   )
 
   @doc """
@@ -402,7 +402,7 @@ defmodule Domain.Config.Definitions do
   """
   defconfig(:cookie_signing_salt, :string,
     sensitive: true,
-    changeset: &Domain.Validator.validate_base64/2
+    changeset: &Domain.Repo.Changeset.validate_base64/2
   )
 
   @doc """
@@ -410,7 +410,7 @@ defmodule Domain.Config.Definitions do
   """
   defconfig(:cookie_encryption_salt, :string,
     sensitive: true,
-    changeset: &Domain.Validator.validate_base64/2
+    changeset: &Domain.Repo.Changeset.validate_base64/2
   )
 
   ##############################################
@@ -484,8 +484,8 @@ defmodule Domain.Config.Definitions do
     sensitive: true,
     changeset: fn changeset, key ->
       changeset
-      |> Domain.Validator.trim_change(key)
-      |> Domain.Validator.validate_email(key)
+      |> Domain.Repo.Changeset.trim_change(key)
+      |> Domain.Repo.Changeset.validate_email(key)
     end
   )
