+++ elixir/apps/domain/test/domain/tokens_test.exs
@@ -17,7 +17,7 @@ defmodule Domain.TokensTest do
     }
   end
 
-  describe "fetch_token_by_id/2" do
+  describe "fetch_token_by_id/3" do
     test "returns error when subject does not have required permissions", %{
       subject: subject
     } do
@@ -55,7 +55,7 @@ defmodule Domain.TokensTest do
     end
   end
 
-  describe "list_tokens_for/1" do
+  describe "list_subject_tokens/2" do
     test "returns current subject's tokens", %{
       account: account,
       identity: identity,
@@ -65,7 +65,7 @@ defmodule Domain.TokensTest do
       Fixtures.Tokens.create_token(account: account)
       Fixtures.Tokens.create_token()
 
-      assert {:ok, tokens, _metadata} = list_tokens_for(subject)
+      assert {:ok, tokens, _metadata} = list_subject_tokens(subject)
       token_ids = Enum.map(tokens, & &1.id)
       assert token.id in token_ids
       assert subject.token_id in token_ids
@@ -77,7 +77,7 @@ defmodule Domain.TokensTest do
     } do
       subject = Fixtures.Auth.remove_permissions(subject)
 
-      assert list_tokens_for(subject) ==
+      assert list_subject_tokens(subject) ==
                {:error,
                 {:unauthorized,
                  reason: :missing_permissions,
@@ -85,7 +85,7 @@ defmodule Domain.TokensTest do
     end
   end
 
-  describe "list_tokens_for/2" do
+  describe "list_tokens_for/3" do
     test "returns tokens of a given actor", %{
       account: account,
       subject: subject
